From nikc --
Using extract_xpath I've discovered that it doesn't handle (at least)
text() nodes and attribute values.

For example, give this:

  <h1>Some text with <em>extra markup</em></h1>

and just wanting the text, the XPath

  extract_xpath:
    title: //h1/descendant::text()

fails.  Or this:

  <meta name="DC.date.modified" content="2006-05-16 15:34:30">

and trying to extract the date:

  extract_xpath:
    date: //meta[@name="DC.date.modified"]/@content

This is because it always calls the as_XML() method on the extracted
data, and neither of those types support that method.